Fix error when used with bookshelf-jsonapi-params
Why:
Because the returned value is `model.fetchAll()`
not the model itself, hence the call to other module
after `model.flexiWhere` yield an error.

This change addresses the need by:
Returning model instead of `model.fetchAll()`.
This actually relevant with the purpose of
this package to only add queries without
executing it. Hence it can be combined
with another bookshelf plugin easily